requirements.txt
torch
torch-fidelity
torchaudio
torchdiffeq
torchrl
torchsde
torchvision
absl-py
accelerate
addict
aiohappyeyeballs
aiohttp
aiosignal
alembic
annotated-types
anthropic
antlr4-python3-runtime
anyio
anykeystore
apex==0.9.10dev
astor
astunparse
async-timeout
attrs
bitsandbytes
blinker
certifi
charset-normalizer
click
cloudpickle
codebleu
codecov
colorama
colorlog
contourpy
coverage
cryptacular
cycler
datasets
defusedxml
deprecated
detectors
dictdot
dill
distro
dm-env
dm-tree
dmlab2d
docker-pycreds
editdistance
einops
etils
eval-type-backport
faiss-cpu
filelock
fire
flake8
flask
flask-sqlalchemy
flask-wtf
flatbuffers
flax
fonttools
frozenlist
fsspec
future
gast
gitdb
gitpython
google-pasta
grpcio
h11
h5py
httpcore
httpx
huggingface-hub
humanize
hupper
hydra-core
idna
imageio
imageio-ffmpeg
immutabledict
importlib-resources
iniconfig
interrogate
ipywidgets
isort
itsdangerous
jax
jaxlib
jinja2
jiter
joblib
jupyterlab-widgets
keras
kiwisolver
lazy-loader
levenshtein
libclang
littleutils
mako
markdown
markdown-it-py
markupsafe
matplotlib
mccabe
mdurl
mistralai
ml-collections
ml-dtypes
mmcls
git+https://github.com/open-mmlab/mmcv.git@d409eedc816fccfb1c8d57e5eed5f03bd075f327
git+https://github.com/Lakonik/mmgeneration.git@500a93e474638c87dcaf8fad94cfbe1ef29acd1f
git+https://github.com/huggingface/diffusers.git@57084dacc5275d7212513b24837b60a28e55603d
mmengine
mmgen
mpmath
msgpack
multidict
multiprocess
namex
networkx
ninja
numpy
oauthlib
ogb
ollama
omegaconf
openai
opencv-python
openreview-py
opt-einsum
optax
optree
optuna
orbax-checkpoint
orjson
outdated
pandas
pastedeploy
pbkdf2
pebble
pillow
plaster
plaster-pastedeploy
pluggy
prettytable
propcache
protobuf
py
pyarrow
pybind11
pycodestyle
pycryptodome
pydantic
pydantic-core
pyflakes
pygame
pyjwt
pylatexenc
pylcs
pymupdf
pyparsing
pypdf2
pyramid
pyramid-mailer
pyspng
pytest
pytest-runner
python-dotenv
python3-openid
pytz
pyyaml
rapidfuzz
reactivex
regex
repoze-sendmail
requests
requests-oauthlib
rich
safetensors
scikit-image
scikit-learn
scipy
seaborn
sentry-sdk
setproctitle
simplejson
smmap
sniffio
sqlalchemy
sympy
tabulate
tensorboard
tensorboard-data-server
tensordict
tensorflow
tensorflow-io-gcs-filesystem
tensorstore
termcolor
threadpoolctl
tifffile
tiktoken
timm
tld
tokenizers
tomli
toolz

tqdm
trampoline
transaction
transformers
translationstring
treescope
typing-inspection
tzdata
urllib3
velruse
venusian
wandb
webob
werkzeug
widgetsnbextension
wilds
wrapt
wtforms
wtforms-recaptcha
xxhash
yapf==0.40.1
yarl
zope-deprecation
zope-interface
zope-sqlalchemy
